requirements.txt
pandas
openpyxl
langchain
langchain-ollama
faiss-cpu
numpy
scikit-learn
langchain-community
transformers 
sentence-transformers
transformers

django
markdown2
mysql-connector-python